Where to paste the files
Paste all the files in website2 folder in the source folder of your respective IDE's. For JetBrains IDE's, Pycharm and Intellij paste them in src folder in any of your projects. If you are using Clion paste website2 folder Directly in the path. You will find the exact "path", in the picture I uploaded above